fix: 명동합정 노선 옵챗 추가 및 reservation card 에도 반영
